fix: move TRT-LLM logprob parity test to the GPU lane
Importing HandlerBase pulls in the native TRT-LLM bindings, which need
libcuda. The shared logprobs module is gpu_0, so the test failed on the
CPU-only lane with an ImportError; its importorskip guard no longer
catches it because the top-level tensorrt_llm package imports without a
GPU. Move the test to the gpu_1 TRT-LLM handler module instead.
