+++ components/src/dynamo/trtllm/tests/test_trtllm_handler_base.py
@@ -6,6 +6,7 @@ import logging
 import re as re_mod
 from copy import deepcopy
 from dataclasses import dataclass
+from types import SimpleNamespace
 from typing import Any
 from unittest import mock
 from unittest.mock import MagicMock
@@ -23,6 +24,7 @@ from tensorrt_llm.executor.request import DEFAULT_REQUEST_PRIORITY
 from tensorrt_llm.llmapi import DisaggregatedParams
 from tensorrt_llm.llmapi.llm import SamplingParams
 
+from dynamo.common.backend.logprobs import extract_from_completion_output
 from dynamo.llm.exceptions import EngineShutdown
 from dynamo.trtllm.constants import DisaggregationMode
 from dynamo.trtllm.health_check import TrtllmHealthCheckPayload
@@ -1632,3 +1634,32 @@ class TestEngineIdMapLogging:
         assert "trtllm_client_id=51420" in maps[0]
         # Aggregated mode carries no cross-phase disagg id.
         assert "disagg_request_id=None" in maps[0]
+
+
+def _logprob(lp: float, rank: int = 1, decoded: str | None = None):
+    return SimpleNamespace(logprob=lp, rank=rank, decoded_token=decoded)
+
+
+def test_trtllm_handler_matches_shared():
+    """HandlerBase must not drift from the shared logprob extractor.
+
+    Lives here rather than beside the other shared-logprob tests in
+    components/src/dynamo/common/backend/tests/test_logprobs.py because
+    importing HandlerBase pulls in the native TRT-LLM bindings, which need a
+    GPU, and that module is gpu_0.
+    """
+    output = SimpleNamespace(
+        token_ids=[11, 12],
+        logprobs=[
+            {11: _logprob(-0.1), 110: _logprob(-1.1)},
+            # Selected token missing -- exercises the fallback flag.
+            {99: _logprob(-9.9)},
+        ],
+    )
+
+    wrapper_lp, wrapper_top = HandlerBase._extract_logprobs(output, 0)
+    direct_lp, direct_top = extract_from_completion_output(
+        output, 0, fallback_to_first_on_missing=True, include_bytes=False
+    )
+    assert wrapper_lp == direct_lp
+    assert wrapper_top == direct_top
